Update and rename MARCELS to MARCELS.py
This is a MARELS prediction based on the article on PitcherList from Jeremey Seigel. Jeremy used R in his article and this is the Python translation to what he wrote. This idea and structure is credited to Jeremy and is no way my intention to claim it is my idea. The purpose of this was to show that Python is able to be used much like R in the baseball analytics space as well as provide a learning material for those interested in Python.