Add files via upload
This script downloads data from pybaseball from yesterday's games. It cleans the data and also changes some column names so that wrangling data is simpler for me (specific names that I use instead of the ones Savant uses). The data is exported to a sql database. I will run this with CRON to automatically run at a certain time each morning.